requirements.txt
Flask==1.1.1
gunicorn==19.9.0
PyYAML==5.1.2
requests==2.22.0
